server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt>=5.0.0",
    "deepface==0.0.93",
    "fastapi[standard]>=0.120.2",
    "msal>=1.34.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.44",
    "tf-keras>=2.20.1",
    "uuid>=1.30",
    "werkzeug>=3.1.3",
]
